Adds a Consul Resolver for Generated Records
This commit adds multiple-resolver support for mesos-dns which allows it
to support multiple backends for publishing discovered records including
the default `builtin` resolver and a new `consul` resolver that
duplicates and improves upon `mesos-consul`
(https://github.com/CiscoCloud/mesos-consul) capabilities.

The primary thrust of this work is to allow us to take advantage of the
excellent records generator already found in `mesos-dns` but add
pluggable options for how those records can be published and/or
consumed.

Features include:

- adding a docker build environment and Makefile to speed and
  standardize testing

- updating the golang version to latest stable

- adding gofmt test targets

- modularizing the resolver to allow multiple backends

- implementing a new consul backend

- formatting timestamps to RFC3339Nano for more reliable parsing

- adding state attributes and task labels to the records generator

Improvements to existing mesos-consul features include:

- adding variable substitution (`{IP}` and `{PORT}`)

- adding fallback options to use the consul catalog if the agent is
  unavailable

- adding task-label-defined consul healthchecks

- adding support to add each slave to a given service via a consul tag
  to allow local access eg, http://slave-01.servicename.service.consul/